Revert "create new submodule"
This reverts commit 32be1d13ecf5cf52ff33b15d982112b6bb06de3e.
